Compile OpenFOAM in 'opt' mode in Dockerfile
There was some issues with the pinned apt versions as well, which have
been resolved.

+++ include/base/FoamInterface.h
@@ -14,7 +14,7 @@ namespace Foam
 class polyPatch;
 template <typename T>
 class List;
-typedef int label;
+typedef long int label;
 typedef List<label> labelList;
 template <typename T>
 class autoPtr;
