+++ include/mesh/Foam2MooseMeshGen.h
@@ -10,7 +10,7 @@ namespace Foam
 {
 template <typename T>
 class IOList;
-typedef int label;
+typedef long int label;
 typedef IOList<label> labelIOList;
 }
